Auto cleanup

+++ src/main/java/de/ellpeck/actuallyadditions/mod/blocks/BlockCoffeeMachine.java
@@ -88,7 +88,7 @@ public class BlockCoffeeMachine extends BlockContainerBase{
 
     @Override
     public void onBlockPlacedBy(World world, BlockPos pos, IBlockState state, EntityLivingBase player, ItemStack stack){
-        int rotation = MathHelper.floor((double)(player.rotationYaw*4.0F/360.0F)+0.5D) & 3;
+        int rotation = MathHelper.floor(player.rotationYaw*4.0F/360.0F+0.5D) & 3;
 
         if(rotation == 0){
             world.setBlockState(pos, this.getStateFromMeta(0), 2);
